+++ src/osc_transformer_based_extractor/fine_tune.py
@@ -54,6 +54,17 @@ def check_output_dir(output_dir):
 
 
 class CustomDataset(Dataset):
+    """
+    Custom dataset class for sequence classification tasks.
+
+    Args:
+        tokenizer (transformers.PreTrainedTokenizer): The tokenizer to use for encoding the data.
+        questions (pd.Series): A pandas series containing the questions.
+        contexts (pd.Series): A pandas series containing the contexts.
+        labels (pd.Series): A pandas series containing the labels.
+        max_length (int): The maximum length of the input sequences.
+    """
+
     def __init__(self, tokenizer, questions, contexts, labels, max_length):
         self.tokenizer = tokenizer
         self.questions = questions
@@ -62,9 +73,19 @@ class CustomDataset(Dataset):
         self.max_length = max_length
 
     def __len__(self):
+        """Return the number of samples in the dataset."""
         return len(self.questions)
 
     def __getitem__(self, idx):
+        """
+        Get the sample at index `idx`.
+
+        Args:
+            idx (int): Index of the sample to retrieve.
+
+        Returns:
+            dict: A dictionary containing input IDs, attention mask, and labels.
+        """
         question = str(self.questions[idx])
         context = str(self.contexts[idx])
         label = self.labels[idx]
@@ -84,6 +105,19 @@ class CustomDataset(Dataset):
 
 
 def fine_tune_model(data_path, model_name, num_labels, max_length, epochs, batch_size, output_dir, save_steps):
+    """
+    Fine-tune a pre-trained model on a custom dataset.
+
+    Args:
+        data_path (str): Path to the CSV file containing the dataset.
+        model_name (str): Name/path of the pre-trained model to use.
+        num_labels (int): Number of labels for the classification task.
+        max_length (int): Maximum length of the input sequences.
+        epochs (int): Number of training epochs.
+        batch_size (int): Batch size for training.
+        output_dir (str): Directory where the model will be saved during training.
+        save_steps (int): Number of steps before saving the model during training.
+    """
     # Load your dataset into a pandas DataFrame
     df = pd.read_csv(data_path)
     df = df[["question", "context", "label"]]
@@ -192,5 +226,4 @@ python fine_tune.py \
   --batch_size 4 \
   --output_dir "./saved_models_during_training" \
   --save_steps 500
-
 '''
