+++ src/osc_transformer_based_extractor/inference.py
@@ -28,6 +28,16 @@ def check_model_and_tokenizer_path(model_path, tokenizer_path):
 
 
 def check_question_context(question, context):
+    """
+    Check if the question and context are valid strings and not empty.
+
+    Args:
+        question (str): The question string.
+        context (str): The context string.
+
+    Raises:
+        ValueError: If the question or context is not a string or is empty.
+    """
     if not isinstance(question, str):
         raise ValueError("Question must be a string.")
 
@@ -91,10 +101,12 @@ if __name__ == "__main__":
 
     print(f"Predicted Label ID: {result}")
 
+'''
+To run the file in CMD:
 
-
-'''python inference.py
-    --question "What is the capital of France?"
-    --context "Paris is the capital of France."
-    --model_path /path/to/model
-    --tokenizer_path /path/to/tokenizer'''
+python inference.py \
+  --question "What is the capital of France?" \
+  --context "Paris is the capital of France." \
+  --model_path /path/to/model \
+  --tokenizer_path /path/to/tokenizer
+'''
