+++ src/package/osc-transformer-based-extractor/osc_transformer_based_extractor/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains CLI script for the OSC Transformer-Based Extractor.
+"""
