+++ src/package/osc-transformer-based-extractor/tests/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains tests for CLI for the OSC Transformer-Based Extractor.
+"""
