+++ src/pytests/src/PYTESTS/testing_data/valid_curator_data.csv
@@ -1,0 +1,2 @@
+context,kpi_id,label
+Lorem ipsum,1,1
