+++ src/pytests/src/PYTESTS/testing_data/valid_kpi_mapping.csv
@@ -1,0 +1,2 @@
+kpi_id,question,sectors,add_year,kpi_category
+1,What is the question?,Sector,2022,Category
